tests: make tests pass
I don't know how some of these tests passed in the first place
and that makes me very anxious.
